requirements.txt
dockerfile==3.2.0
bashlex==0.18
PyYAML~=6.0
ansible==8.0.0
regex~=2022.10.31
oyaml~=1.0
antlr4-python3-runtime~=4.11.1
tabulate==0.9.0
tqdm~=4.64.1
munch~=2.5.0
parts==1.5.2
validators==0.20.0
Jinja2~=3.1.2